pc/requirements.txt
pyserial
numpy
mss
fast_colorthief